Use Playnite time converter

+++ LiveChartsCommon/CustomerForTime.cs
@@ -5,7 +5,7 @@ namespace PluginCommon.LiveChartsCommon
     public class CustomerForTime
     {
         public string Name { get; set; }
-        public double Values { get; set; }
+        public long Values { get; set; }
         public string ValuesFormat => (int)TimeSpan.FromSeconds(Values).TotalHours + "h " + TimeSpan.FromSeconds(Values).ToString(@"mm") + "min";
     }
 }
